+++ src/test/java/io/github/co_mmer/aaamockmvc/ej/scenario/res/data/ScenarioDataActPost.java
@@ -12,7 +12,7 @@ import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_HEADER_MAP_1_2;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_HEADER_VALUE_1;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_HEADER_VALUE_2;
-import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestObject.TEST_REQUEST_CONFIG;
+import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestObject.TEST_REQUEST_BEAN;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestParameter.TEST_PARAM_KEY_1;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestParameter.TEST_PARAM_KEY_2;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestParameter.TEST_PARAM_KEY_VALUE_MAP_1_2;
@@ -38,7 +38,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class ScenarioDataActPost {
 
-  private static final TestRequestPost POST = new TestRequestPost(TEST_REQUEST_CONFIG);
+  private static final TestRequestPost POST = new TestRequestPost(TEST_REQUEST_BEAN);
 
   public static final TestAct1 POST_ARRANGE__URL = POST.arrange().arrangeUrl(TEST_PATH).act();