+++ src/test/java/io/github/co_mmer/aaamockmvc/ej/test/web/arrange/get/head/TestArrangeGetHeaderImplTest.java
@@ -66,8 +66,8 @@ class TestArrangeGetHeaderImplTest {
 
   @Test
   @SuppressWarnings("ConstantConditions")
-  void GIVEN_null_WHEN_arrangeAccept_THEN_throw_NullPointerException() {
-    assertThrows(NullPointerException.class, () -> this.impl.arrangeAccept((MediaType) null));
+  void GIVEN_null_WHEN_arrangeAccept_THEN_throw_IllegalArgumentException() {
+    assertThrows(IllegalArgumentException.class, () -> this.impl.arrangeAccept((MediaType) null));
   }
 
   @Test
@@ -103,8 +103,9 @@ class TestArrangeGetHeaderImplTest {
 
   @Test
   @SuppressWarnings("ConstantConditions")
-  void GIVEN_null_WHEN_arrangeContentType_THEN_throw_NullPointerException() {
-    assertThrows(NullPointerException.class, () -> this.impl.arrangeContentType((MediaType) null));
+  void GIVEN_null_WHEN_arrangeContentType_THEN_throw_IllegalArgumentException() {
+    assertThrows(
+        IllegalArgumentException.class, () -> this.impl.arrangeContentType((MediaType) null));
   }
 
   @Test