+++ src/test/java/io/github/co_mmer/aaamockmvc/ej/testdata/testutil/TestJsonProcessingExceptionWrapper.java
@@ -1,0 +1,10 @@
+package io.github.co_mmer.aaamockmvc.ej.testdata.testutil;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+public class TestJsonProcessingExceptionWrapper extends JsonProcessingException {
+
+  public TestJsonProcessingExceptionWrapper() {
+    super("test");
+  }
+}